[DSPlayer] updated subtitle project (now builds with SSE2 optimization)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@30563 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/DSPlayer/Subtitles/subpic/CoordGeom.h
@@ -62,7 +62,7 @@ public:
   Vector Refract2(Vector& n, float nFrom, float nTo, float* nOut = NULL);
 
   Vector operator - ();
-  float& operator [] (int i);
+  float& operator [] (size_t i);
 
   float operator | (Vector& v); // dot
   Vector operator % (Vector& v); // cross
