+++ xbmc/cores/DSPlayer/Subtitles/subpic/DX9SubPic.cpp
@@ -257,7 +257,7 @@ STDMETHODIMP CDX9SubPic::AlphaBlt(RECT* pSrc, RECT* pDst, SubPicDesc* pTarget)
       {(float)dst.right, (float)dst.bottom, 0.5f, 2.0f, (float)src.right / w, (float)src.bottom / h},
     };
 /*
-    for(int i = 0; i < countof(pVertices); i++)
+	for(ptrdiff_t i = 0; i < countof(pVertices); i++)
     {
       pVertices[i].x -= 0.5;
       pVertices[i].y -= 0.5;
